Complete Task 53: Grid Expansion to 64 Steps

This merge brings in the 64-step grid expansion (Phase 5 of the epic):
- 4 commits with implementation, fixes, and documentation
- All 373 tests passing
- Code review complete (6 critical issues fixed)
- Backlog tasks reviewed for compatibility

Epic status: All phases (1-5) complete. Task 65 deferred to backlog.

+++ src/Engine/DriftControl.cpp
@@ -67,9 +67,9 @@ float GetStepStability(int step, int patternLength)
     return kStabilityOffbeat;
 }
 
-uint32_t GetStabilityMask(int patternLength, float stabilityThreshold)
+uint64_t GetStabilityMask(int patternLength, float stabilityThreshold)
 {
-    uint32_t mask = 0;
+    uint64_t mask = 0;
 
     int maxSteps = std::min(patternLength, static_cast<int>(kMaxSteps));
 
@@ -78,7 +78,7 @@ uint32_t GetStabilityMask(int patternLength, float stabilityThreshold)
         float stability = GetStepStability(step, patternLength);
         if (stability >= stabilityThreshold)
         {
-            mask |= (1U << step);
+            mask |= (1ULL << step);
         }
     }
